TNC-Analyzer/python_files/requirements.txt
pywin32
Pillow
pytesseract
google-genai
comtypes
dotenv
keyboard